web-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
ultralytics==8.0.200
torch>=2.0.0
Pillow==10.0.1
numpy>=1.24.0
opencv-python>=4.8.0
werkzeug>=2.3.7
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
watchdog>=2.0.0

